fix(desktop): resolve 3 Kilo findings on TodoApp — Share panel note, load error state, duplicate test suites
- WARNING: Added explicit comment documenting that standalone TodoApp does not yet carry Share/members panel or per-entry revision history from the old DocsApp list variant — tracked as follow-up work
- SUGGESTION: Added loadError state + retry button in loadLists — fetch failures now show an error alert instead of infinite spinner
- SUGGESTION: Consolidated Todo tests into __tests__/NotesApp.test.tsx, removed duplicate describe("TodoApp") from NotesApp.test.tsx

Docs-Reviewed: README already lists both Notes and Todo under platform apps; standalone TodoApp.tsx extraction just makes the component boundary match the docs
